Improve funtionality of maintenance

+++ technical_assistance/views.py
@@ -16,26 +16,29 @@ class MaintenanceListView(ListView):
 		context = super(MaintenanceListView, self).get_context_data(**kwargs)
 
 		allocations = Allocation.objects.filter(is_active=True)
-		devices = Device.objects.filter(model__type__is_part=True).exclude(parts__pk__gt=0)						
+		devices = Device.objects.filter(model__type__is_part=True).exclude(parts__pk__gt=0)
 
 		context['allocations'] = allocations
-		context['devices'] = devices		
+		context['devices'] = devices
 		return context
 
 	def post(self, request, *args, **kwargs):
-		if request.is_ajax():			
-			maintenance_modelform = modelform_factory(Maintenance, fields=('date', 'problem', 'solution', 'device', ))					
+		if request.is_ajax():
+			specifications = request.POST.get('specifications', None)
+			maintenance_modelform = modelform_factory(Maintenance, fields=('date', 'problem', 'solution', 'device', ))
     		maintenance_form = maintenance_modelform(request.POST)
-    		if maintenance_form.is_valid():
-    			object = maintenance_form.save()
-    			data = model_to_dict(object)
-    			parts = request.POST['parts'].split(',')
-    			for id in parts:
-    				device = Device.objects.get(pk=id)
-    				part = Parts(is_active=True, part=device, maintenance=object)
-    				part.save()    				
-    			return JsonResponse(data)
-    		return JsonResponse({}, status=400)				
+    		if maintenance_form.is_valid() and specifications is not None:
+				object = maintenance_form.save()
+				object.device.specifications = specifications
+				object.device.save()
+				data = model_to_dict(object)
+				parts = request.POST['parts'].split(',') if request.POST['parts'] != '' else []
+				for id in parts:
+					device = Device.objects.get(pk=id)
+					part = Parts(is_active=True, part=device, maintenance=object)
+					part.save()
+				return JsonResponse(data)
+    		return JsonResponse({}, status=400)
 
 class MaintenanceDetailView(DetailView):
 	model = Maintenance
@@ -48,26 +51,29 @@ class MaintenanceDetailView(DetailView):
 		    for part in Parts.objects.filter(maintenance = self.object):
 		    	item = model_to_dict(part)
 		    	item['name'] = str(part)
-		    	parts.append(item)		    
-		    data['parts'] = parts   
+		    	parts.append(item)
+		    data['parts'] = parts
 		    return JsonResponse(data, safe=False)
 
 	def post(self, request, *args, **kwargs):
-		if request.is_ajax():				
+		if request.is_ajax():
+			specifications = request.POST.get('specifications', None)
 			self.object = self.get_object()
-			maintenance_modelform = modelform_factory(Maintenance, fields=('date', 'problem', 'solution', 'device', ))					
-    		maintenance_form = maintenance_modelform(request.POST, instance=self.object)
-    		if maintenance_form.is_valid():
-    			maintenance_form.save()
-    			self.object.parts_set.all().delete()
-    			data = model_to_dict(self.object)
-    			parts = request.POST['parts'].split(',')
-    			for id in parts:
-    				device = Device.objects.get(pk=id)
-    				part = Parts(is_active=True, part=device, maintenance=self.object)
-    				part.save()
-    			return JsonResponse(data)
-    		return JsonResponse({}, status=400)						
+			maintenance_modelform = modelform_factory(Maintenance, fields=('date', 'problem', 'solution', 'device', ))
+			maintenance_form = maintenance_modelform(request.POST, instance=self.object)
+			if maintenance_form.is_valid() and specifications is not None:
+				maintenance_form.save()
+				self.object.device.specifications = specifications
+				self.object.device.save()
+				self.object.parts_set.all().delete()
+				data = model_to_dict(self.object)
+				parts = request.POST['parts'].split(',') if request.POST['parts'] != '' else []
+				for id in parts:
+					device = Device.objects.get(pk=id)
+					part = Parts(is_active=True, part=device, maintenance=self.object)
+					part.save()
+				return JsonResponse(data)
+			return JsonResponse({}, status=400)
 
 	def delete(self, request, *args, **kwargs):
 		if request.is_ajax():
